src/requirements/base.txt
coverage==7.4.0
flake8==7.0.0
psycopg-binary==3.1.17
python-dotenv==1.0.0
